Updated code with reusability and solved Combobox error.
