+++ src/pages/dashboard/Dashboard.tsx
@@ -1,19 +1,7 @@
+import AddLotDialog from "@/components/AddLotDialog";
+import CustomTable from "@/components/CustomTable";
+import PageHeader from "@/components/PageHeader";
 import { Button } from "@/components/ui/button";
-import {
-    Command,
-    CommandEmpty,
-    CommandGroup,
-    CommandInput,
-    CommandItem,
-} from "@/components/ui/command";
-import {
-    Dialog,
-    DialogContent,
-    DialogFooter,
-    DialogHeader,
-    DialogTitle,
-    DialogTrigger
-} from "@/components/ui/dialog";
 import {
     DropdownMenu,
     DropdownMenuContent,
@@ -22,207 +10,51 @@ import {
     DropdownMenuSeparator,
     DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import {
-    Popover,
-    PopoverContent,
-    PopoverTrigger,
-} from "@/components/ui/popover";
-import {
-    Table,
-    TableBody,
-    TableCell,
-    TableHead,
-    TableHeader,
-    TableRow
-} from "@/components/ui/table";
-import { cn } from "@/lib/utils";
-import { Check, ChevronsUpDown, MoreHorizontal } from "lucide-react";
-import * as React from "react";
-import { FaSearch, FaUser } from 'react-icons/fa';
 
-
-const frameworks = [
-    {
-        value: "next.js",
-        label: "Next.js",
-    },
-    {
-        value: "sveltekit",
-        label: "SvelteKit",
-    },
-    {
-        value: "nuxt.js",
-        label: "Nuxt.js",
-    },
-    {
-        value: "remix",
-        label: "Remix",
-    },
-    {
-        value: "astro",
-        label: "Astro",
-    },
-]
+import { MoreHorizontal } from "lucide-react";
 
 
 function Dashboard() {
-    const [open, setOpen] = React.useState(false)
-    const [value, setValue] = React.useState("")
+
+    const tableHeaders = ["PRODUCT", "VENDOR", "AGREEMENTTYPE", "DOCUMENT", ""];
+    const tableData = [
+        ["Product 1", "pradip", "My order", (
+            <DropdownMenu>
+                <DropdownMenuTrigger asChild>
+                    <Button variant="ghost" className="h-8 w-8 p-0">
+                        <span className="sr-only">Open menu</span>
+                        <MoreHorizontal className="h-4 w-4" />
+                    </Button>
+                </DropdownMenuTrigger>
+                <DropdownMenuContent align="end">
+                    <DropdownMenuLabel>Home</DropdownMenuLabel>
+                    <DropdownMenuItem>Admin</DropdownMenuItem>
+                    <DropdownMenuSeparator />
+                    <DropdownMenuItem>User Profile</DropdownMenuItem>
+                    <DropdownMenuItem>Organization</DropdownMenuItem>
+                </DropdownMenuContent>
+            </DropdownMenu>
+        ), (
+                <Button className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-1 px-2 rounded">
+                    View
+                </Button>
+            )]
+    ];
+
     return (
         <div>
-            <div className="flex items-center justify-between bg-gradient-to-r from-sky-200 to-blue-700 py-10">
-                <h1 className="text-lg font-bold text-white px-4">DASHBOARD</h1>
-                <div className="flex items-center space-x-2">
-                    <div className="relative rounded-full bg-white">
-                        <FaSearch className="absolute left-2 top-1/2 transform -translate-y-1/2 ml-2 text-gray-500" />
-                        <Input className="pl-8 bg-transparent rounded-full focus:outline-none" placeholder="Search..." />
-                    </div>
-                    <a href="/profile" className="px-2">
-                        <FaUser className="text-white text-xl cursor-pointer" />
-                    </a>
-                </div>
+            <div>
+                <PageHeader title="DASHBOARD" />
             </div>
             <div className="flex justify-between items-center mb-4 py-5 px-10">
                 <b>Add Lot</b>
                 <div>
-                    <Dialog>
-                        <DialogTrigger asChild>
-                            <Button variant="outline" className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-1 rounded px-5">Add Lot</Button>
-                        </DialogTrigger>
-                        <DialogContent className="sm:max-w-[425px] bg-white">
-                            <DialogHeader>
-                                <DialogTitle>Create Lot</DialogTitle>
-                            </DialogHeader>
-                            <div className="grid gap-4 py-4">
-                                <div className="grid grid-cols-4 items-center gap-4">
-                                    <Label htmlFor="name" className="text-right">
-                                        Shipper Id
-                                    </Label>
-                                    <Popover open={open} onOpenChange={setOpen}>
-                                        <PopoverTrigger asChild>
-                                            <Button
-                                                variant="outline"
-                                                role="combobox"
-                                                aria-expanded={open}
-                                                className="w-[200px] justify-between"
-                                            >
-                                                {value
-                                                    ? frameworks.find((framework) => framework.value === value)?.label
-                                                    : "Select framework..."}
-                                                <ChevronsUpDown className="ml-2 h-4 w-4 shrink-0 opacity-50" />
-                                            </Button>
-                                        </PopoverTrigger>
-                                        <PopoverContent className="w-[200px] p-0">
-                                            <Command>
-                                                <CommandInput placeholder="Search framework..." />
-                                                <CommandEmpty>No framework found.</CommandEmpty>
-                                                <CommandGroup>
-                                                    {frameworks.map((framework) => (
-                                                        <CommandItem
-                                                            key={framework.value}
-                                                            value={framework.value}
-                                                            onSelect={(currentValue) => {
-                                                                setValue(currentValue === value ? "" : currentValue)
-                                                                setOpen(false)
-                                                            }}
-                                                        >
-                                                            <Check
-                                                                className={cn(
-                                                                    "mr-2 h-4 w-4",
-                                                                    value === framework.value ? "opacity-100" : "opacity-0"
-                                                                )}
-                                                            />
-                                                            {framework.label}
-                                                        </CommandItem>
-                                                    ))}
-                                                </CommandGroup>
-                                            </Command>
-
-                                        </PopoverContent>
-                                    </Popover>
-                                </div>
-                                <div className="grid grid-cols-4 items-center gap-4">
-                                    <Label htmlFor="name" className="text-right">
-                                        Bank Id
-                                    </Label>
-                                    <Input
-                                        id="name"
-                                        defaultValue="Pedro Duarte"
-                                        className="col-span-3"
-                                    />
-                                </div>
-                                <div className="grid grid-cols-4 items-center gap-4">
-                                    <Label htmlFor="name" className="text-right">
-                                        Wholesaler Id
-                                    </Label>
-                                    <Input
-                                        id="name"
-                                        defaultValue="Pedro Duarte"
-                                        className="col-span-3"
-                                    />
-                                </div>
-                                <div className="grid grid-cols-4 items-center gap-4">
-                                    <Label htmlFor="username" className="text-right">
-                                        Select Excel
-                                    </Label>
-                                    <Input
-                                        id="username"
-                                        defaultValue="@peduarte"
-                                        className="col-span-3"
-                                    />
-                                </div>
-                            </div>
-                            <DialogFooter>
-                                <Button type="submit">Upload</Button>
-                                <Button type="submit">Cancle</Button>
-                            </DialogFooter>
-                        </DialogContent>
-                    </Dialog>
-
+                    <AddLotDialog label="Add Lot" />
                 </div>
             </div>
-            <Table>
-                <TableHeader className="bg-slate-200">
-                    <TableRow>
-                        <TableHead>PRODUCT</TableHead>
-                        <TableHead>VENDOR</TableHead>
-                        <TableHead>AGREEMENTTYPE</TableHead>
-                        <TableHead>DOCUMENT</TableHead>
-                        <TableHead></TableHead>
-                    </TableRow>
-                </TableHeader>
-                <TableBody>
-                    <TableRow>
-                        <TableCell>Product 1</TableCell>
-                        <TableCell>pradip</TableCell>
-                        <TableCell>My order</TableCell>
-                        <TableCell>
-                            <DropdownMenu>
-                                <DropdownMenuTrigger asChild>
-                                    <Button variant="ghost" className="h-8 w-8 p-0">
-                                        <span className="sr-only">Open menu</span>
-                                        <MoreHorizontal className="h-4 w-4" />
-                                    </Button>
-                                </DropdownMenuTrigger>
-                                <DropdownMenuContent align="end">
-                                    <DropdownMenuLabel>Home</DropdownMenuLabel>
-                                    <DropdownMenuItem>Admin</DropdownMenuItem>
-                                    <DropdownMenuSeparator />
-                                    <DropdownMenuItem>User Profile</DropdownMenuItem>
-                                    <DropdownMenuItem>Organization</DropdownMenuItem>
-                                </DropdownMenuContent>
-                            </DropdownMenu>
-                        </TableCell>
-                        <TableCell>
-                            <Button className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-1 px-2 rounded">
-                                View
-                            </Button>
-                        </TableCell>
-                    </TableRow>
-                </TableBody>
-            </Table>
+            <div>
+                <CustomTable headers={tableHeaders} data={tableData} />
+            </div>
         </div>
     );
 }
